// cypress/pages/enums/selectors.ts
export enum SELECTORS {
  //HOME PAGE
  APP_LOGO = '[id="img-xtr-logo"]',
  HEADER = '[id="lbl-header"]',
  FILTER_SEDAN = '[id="btn-sedan"]',
  FILTER_SUV = '[id="btn-suv"]',
  FILTER_HATCHBACK = '[id="btn-hatchback"]',
  FILTER_COUPE = '[id="btn-coupe"]',
  // Featured Cars section
  FILTER_CAR = '[id="btn-filter"]',
  VIEW_ALL_BTN = '[id="btn-view-all"]',
  // Cars section
  CAR_CARD = '[id="car-description-card"]',
  CAR_PREVIEW_IMG = '[id="lbl-car-image"]',
  CAR_NAME = '[id="lbl-car-name"]',
  CAR_PRICE = '[id="lbl-price"]',
  ARROW_LINK = '[id="btn-details"]',
  // Footer section
  CONTACT_SECTION = '[id="section-contact"]',
  ADDRESS_SECTION = '[id="section-address"]',
  LEGAL_SECTION = '[id="section-legal"]',
  SOCIAL_SECTION = '[id="section-social"]',

  // CAR LISTING PAGE
  RESET_FILTER = '[id="btn-reset"]',

  // CAR DETAILS PAGE
  CAR_IMAGE = '[id="img-car"]',
  CAR_FUEL = '[id="section-fuel"]',
  CAR_ENGINE = '[id="section-engine"]',
  CAR_TORQUE = '[id="section-torque"]',
  CAR_ACCELERATION = '[id="section-accel"]',
  CAR_TOP_SPEED = '[id="section-top-spd"]',
  CAR_VARIANTS = '[id="section-variants"]',
  CAR_EXTERIORS = '[id="section-exteriors"]',
  CAR_INTERIORS = '[id="section-interior"]',
  CAR_COST = '[id="lbl-car-cost"]',
  BOOK_NOW = '[id="btn-book-now"]',

  //CAR BOOKING PAGE
  CAR_SUMMARY = '[id="section-summary"]',
  CAR_BOOKING_NAME = '[id="lbl-car-name"]',
  FUEL_TYPE = '[id="lbl-fuel-type"]',
  ENGINE_CC = '[id="lbl-engine"]',
  VIEW_CAR_DETAILS = '[id="btn-view-details"]',
  USER_NAME_INPUT = '[id="txt-name"]',
  USER_CONTACT_INPUT = '[id="txt-contact"]',
  USER_CITY_SELECT = '[id="lst-city"]',
  USER_CITY_OPTION = '[id="opt-city"]',
  TERMS_CONDITIONS_CHECKBOX = '[id="chk-box"]',
  SUBMIT_BTN = '[id="btn-submit"]',

  //BOOKING CONFIRMATION PAGE
  CONFIRM_BOOKING = '[id="lbl-cnf"]',
  DOWNLOAD_BTN = '[id="btn-download"]',

  //TODO use the following commented ids
  //to write further test cases to check other functionalities

  // ERR_NAME_INPUT="[id='err-txt-name']",
  // ERR_CONTACT_INPUT="[id='err-txt-contact']",
  // ERR_CITY_SELECT="[id='err-lst-city']",
  // ERR_TERMS_CONDITIONS_CHECKBOX="[id='err-chk-box']",
}